+++ docs/conf.py
@@ -14,6 +14,8 @@
 import os
 import sys
 
+# abspath because this could be __main__, in which case it may not
+# have an absolute __file__
 __HERE__ = os.path.dirname(os.path.abspath(__file__))
 
 on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
@@ -29,7 +31,7 @@ if not on_rtd:  # only import and set the theme if we're building docs locally
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
-sys.path.insert(0, os.path.abspath(os.path.join('..')))
+sys.path.insert(0, os.path.abspath('..'))
 
 # -- General configuration ------------------------------------------------
 
@@ -71,9 +73,9 @@ copyright = '2015-2017 — Mozilla Services'
 # built documents.
 #
 # The short X.Y version.
-version = '8.0'
+version = '8.1'
 # The full version, including alpha/beta/rc tags.
-release = '8.0.0'
+release = '8.1.5'
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.